backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
sqlalchemy==2.0.23
alembic==1.13.1
# Optional: For authentication (uncomment when needed)
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4
# pydantic-settings==2.1.0